Add code to hash and upload directories
The client is now capable of hashing a directory and uploading it to the
CAS using either BLAKE3ZCC or SHA256. It will only upload missing blobs
from the directory. It will create batches of FindMissingBlobs calls
before uploading in order to reduce the number of calls made to the
server.